refactor(channels/discord): P5 — extract slash.rs (registration + reconcile)
Move the slash-command subsystem (skill->spec mapping, registration body,
ownership-shape + change-detection projection, DISCORD_API_BASE, rate_limit_
deadline, reconcile_slash_commands — incl. the #7784 ReconcileOutcome/Retry-After
durability) into discord/slash.rs. discord_slash_specs_from_skills keeps its
public path for the orchestrator resolver; ReconcileOutcome lives in types.rs;
discord_slash_state stays a sibling. Behaviour-neutral: 161 discord tests pass.
